+++ R/utils.R
@@ -149,6 +149,7 @@ p <- function(...) paste(sep="", collapse="", ...)
 
 vcapply <- function(...) vapply(..., FUN.VALUE = character(1))
 vlapply <- function(...) vapply(..., FUN.VALUE = logical(1))
+vnapply <- function(...) vapply(..., FUN.VALUE = numeric(1))
 
 iSFALSE <- function(x) {
   is.logical(x) && length(x) == 1L && !is.na(x) && !x